+++ R/options.R
@@ -32,7 +32,9 @@ options::define_option(
 #' @eval options::as_roxygen_docs()
 NULL
 
+#' @title Option parameters
 #' @eval options::as_params()
 #' @name option_params
+#' @keywords internal
 #'
 NULL