+++ NEWS.md
@@ -1,4 +1,14 @@
-# MOSuite (development version)
+# MOSuite development version
+
+- New function `write_multiOmicDataSet_properties()`. (#173, @kelly-sovacool)
+  - Extracts all the properties from a multiOmicDataSet and writes any data frames as csv files, other objects are written as rds files.
+- Bug fixes: (#174, @kelly-sovacool)
+  - Fixed bugs in `plot_volcano_summary()`, `plot_volcano_enhanced()`, and `plot_pca_3d()` when used with multiOmicDataSet objects.
+  - Fixed bug in `filter_diff()` when `filtering_mode = "all"` that was causing plot rendering errors.
+  - Fixed `plot_pca_2d()` to save plots to disk correctly.
+- Improvements for use with Galaxy. (#168, #170, #171, #174, @kelly-sovacool)
+
+# MOSuite 0.2.1
 
 - A docker image is now available. (#134)
   - <https://hub.docker.com/r/nciccbr/mosuite>